M0127

+++ reserve-service/src/main/java/com/bit/docker/reserveservice/api/ReservationController.java
@@ -14,10 +14,10 @@ public class ReservationController {
     }
 
     @PostMapping
-    public Long reserve(
-            @RequestParam Long userId,
-            @RequestParam Long concertId,
-            @RequestParam Long seatId
+    public int reserve(
+            @RequestParam int userId,
+            @RequestParam int concertId,
+            @RequestParam int seatId
     ) {
         return reservationCommandService
                 .reserve(userId, concertId, seatId);
